+++ tests/integ/modin/series/test_to_dynamic_table.py
@@ -3,10 +3,10 @@
 #
 
 import modin.pandas as pd
-import pandas as native_pd
 import pytest
 
 import snowflake.snowpark.modin.plugin  # noqa: F401
+from snowflake.snowpark._internal.utils import TempObjectType
 from snowflake.snowpark.exceptions import SnowparkSQLException
 from snowflake.snowpark.session import Session
 from tests.integ.modin.utils import BASIC_TYPE_DATA1, BASIC_TYPE_DATA2
@@ -14,31 +14,8 @@ from tests.integ.utils.sql_counter import sql_count_checker
 from tests.utils import Utils
 
 
-@pytest.fixture(scope="function")
-def native_pandas_ser_basic():
-    native_ser = native_pd.Series([4, 7, 4, 2], name="A")
-    return native_ser
-
-
-@sql_count_checker(query_count=2)
-def test_create_or_replace_view_basic(session, native_pandas_ser_basic) -> None:
-    view_name = Utils.random_view_name()
-    try:
-        snow_series = pd.Series(native_pandas_ser_basic)
-
-        assert (
-            "successfully created"
-            in snow_series.create_or_replace_view(name=view_name)[0]["status"]
-        )
-    finally:
-        Utils.drop_view(session, view_name)
-
-
-@sql_count_checker(query_count=6)
-def test_create_or_replace_view_multiple_sessions_enforce_ordering_raises(
-    session,
-    db_parameters,
-) -> None:
+@sql_count_checker(query_count=5)
+def test_to_dynamic_table_enforce_ordering_raises(session) -> None:
     try:
         # create table
         table_name = Utils.random_table_name()
@@ -51,33 +28,64 @@ def test_create_or_replace_view_multiple_sessions_enforce_ordering_raises(
             f"(((SELECT * FROM {table_name})))", enforce_ordering=True
         ).iloc[:, 0]
 
-        # create view
-        view_name = Utils.random_view_name()
-        assert (
-            "successfully created"
-            in snow_series.create_or_replace_view(name=view_name)[0]["status"]
+        # creating dynamic_table fails when enforce_ordering is enabled
+        # because it cannot depend on a temp table
+        dynamic_table_name = Utils.random_name_for_temp_object(
+            TempObjectType.DYNAMIC_TABLE
         )
-
-        # another session
-        new_session = Session.builder.configs(db_parameters).create()
-        pd.session = new_session
-
-        # accessing the created view in another session fails when enforce_ordering is enabled
         with pytest.raises(
             SnowparkSQLException,
-            match="Object 'VIEW_NAME' does not exist or not authorized",
+            match="Dynamic Tables cannot depend on a temporary object",
         ):
-            new_session.sql("select * from view_name").collect()
-        new_session.close()
-
+            snow_series.to_dynamic_table(
+                name=dynamic_table_name,
+                warehouse=session.get_current_warehouse(),
+                lag="1000 minutes",
+            )
     finally:
         # cleanup
-        Utils.drop_view(session, view_name)
+        Utils.drop_dynamic_table(session, dynamic_table_name)
+        Utils.drop_table(session, table_name)
+
+
+@sql_count_checker(query_count=5)
+def test_to_dynamic_table_no_enforce_ordering(session) -> None:
+    try:
+        # create table
+        table_name = Utils.random_table_name()
+        session.create_dataframe(
+            [BASIC_TYPE_DATA1, BASIC_TYPE_DATA2]
+        ).write.save_as_table(table_name)
+
+        # create series with enforce_ordering disabled
+        snow_series = pd.read_snowflake(
+            f"(((SELECT * FROM {table_name})))", enforce_ordering=False
+        ).iloc[:, 0]
+
+        # creating dynamic_table succeeds when enforce_ordering is disabled
+        dynamic_table_name = Utils.random_name_for_temp_object(
+            TempObjectType.DYNAMIC_TABLE
+        )
+        assert (
+            "successfully created"
+            in snow_series.to_dynamic_table(
+                name=dynamic_table_name,
+                warehouse=session.get_current_warehouse(),
+                lag="1000 minutes",
+            )[0]["status"]
+        )
+
+        # accessing the created dynamic_table in the same session also succeeds
+        res = session.sql(f"select * from {dynamic_table_name}").collect()
+        assert len(res) == 2
+    finally:
+        # cleanup
+        Utils.drop_dynamic_table(session, dynamic_table_name)
         Utils.drop_table(session, table_name)
 
 
 @sql_count_checker(query_count=4)
-def test_create_or_replace_view_multiple_sessions_no_enforce_ordering(
+def test_to_dynamic_table_multiple_sessions_no_enforce_ordering(
     session,
     db_parameters,
 ) -> None:
@@ -93,25 +101,30 @@ def test_create_or_replace_view_multiple_sessions_no_enforce_ordering(
             f"(((SELECT * FROM {table_name})))", enforce_ordering=False
         ).iloc[:, 0]
 
-        # create view
-        view_name = Utils.random_view_name()
+        # creating dynamic_table succeeds when enforce_ordering is disabled
+        dynamic_table_name = Utils.random_name_for_temp_object(
+            TempObjectType.DYNAMIC_TABLE
+        )
         assert (
             "successfully created"
-            in snow_series.create_or_replace_view(name=view_name)[0]["status"]
+            in snow_series.to_dynamic_table(
+                name=dynamic_table_name,
+                warehouse=session.get_current_warehouse(),
+                lag="1000 minutes",
+            )[0]["status"]
         )
 
         # another session
         new_session = Session.builder.configs(db_parameters).create()
         pd.session = new_session
 
-        # accessing the created view in another session succeeds when enforce_ordering is disabled
-        res = new_session.sql(f"select * from {view_name}").collect()
+        # accessing the created dynamic_table in another session also succeeds
+        res = new_session.sql(f"select * from {dynamic_table_name}").collect()
         assert len(res) == 2
         new_session.close()
-
     finally:
         # cleanup
-        Utils.drop_view(session, view_name)
+        Utils.drop_dynamic_table(session, dynamic_table_name)
         Utils.drop_table(session, table_name)
         pd.session = session
 
@@ -119,7 +132,7 @@ def test_create_or_replace_view_multiple_sessions_no_enforce_ordering(
 @pytest.mark.parametrize("index", [True, False])
 @pytest.mark.parametrize("index_labels", [None, ["my_index"]])
 @sql_count_checker(query_count=4)
-def test_create_or_replace_view_index(session, index, index_labels):
+def test_to_dynamic_table_index(session, index, index_labels):
     try:
         # create table
         table_name = Utils.random_table_name()
@@ -132,9 +145,15 @@ def test_create_or_replace_view_index(session, index, index_labels):
             f"(((SELECT * FROM {table_name})))", enforce_ordering=False
         ).iloc[:, 0]
 
-        view_name = Utils.random_view_name()
-        snow_series.create_or_replace_view(
-            name=view_name, index=index, index_label=index_labels
+        dynamic_table_name = Utils.random_name_for_temp_object(
+            TempObjectType.DYNAMIC_TABLE
+        )
+        snow_series.to_dynamic_table(
+            name=dynamic_table_name,
+            warehouse=session.get_current_warehouse(),
+            lag="1000 minutes",
+            index=index,
+            index_label=index_labels,
         )
         expected_columns = []
         if index:
@@ -148,18 +167,18 @@ def test_create_or_replace_view_index(session, index, index_labels):
 
         # verify columns
         actual = pd.read_snowflake(
-            view_name,
+            dynamic_table_name,
             enforce_ordering=False,
         ).columns
         assert actual.tolist() == expected_columns
     finally:
         # cleanup
-        Utils.drop_view(session, view_name)
+        Utils.drop_dynamic_table(session, dynamic_table_name)
         Utils.drop_table(session, table_name)
 
 
 @sql_count_checker(query_count=4)
-def test_create_or_replace_view_multiindex(session):
+def test_to_dynamic_table_multiindex(session):
     try:
         # create table
         table_name = Utils.random_table_name()
@@ -178,15 +197,19 @@ def test_create_or_replace_view_multiindex(session):
         # create series
         snow_series = snow_dataframe.iloc[:, 0]
 
-        view_name = Utils.random_view_name()
-        snow_series.create_or_replace_view(
-            name=view_name,
+        dynamic_table_name = Utils.random_name_for_temp_object(
+            TempObjectType.DYNAMIC_TABLE
+        )
+        snow_series.to_dynamic_table(
+            name=dynamic_table_name,
+            warehouse=session.get_current_warehouse(),
+            lag="1000 minutes",
             index=True,
         )
 
         # verify columns
         actual = pd.read_snowflake(
-            view_name,
+            dynamic_table_name,
             enforce_ordering=False,
         ).columns
         assert actual.tolist() == ["_1", "_2", "_3"]
@@ -194,10 +217,14 @@ def test_create_or_replace_view_multiindex(session):
         with pytest.raises(
             ValueError, match="Length of 'index_label' should match number of levels"
         ):
-            snow_series.create_or_replace_view(
-                name=view_name, index=True, index_label=["a"]
+            snow_series.to_dynamic_table(
+                name=dynamic_table_name,
+                warehouse=session.get_current_warehouse(),
+                lag="1000 minutes",
+                index=True,
+                index_label=["a"],
             )
     finally:
         # cleanup
-        Utils.drop_view(session, view_name)
+        Utils.drop_dynamic_table(session, dynamic_table_name)
         Utils.drop_table(session, table_name)